Move waitForFullySynchronized and use RunningNode
The contained TODO can be addressed in a next step now.
