+++ hydra-cluster/test/Test/DirectChainSpec.hs
@@ -7,7 +7,6 @@ import Test.Hydra.Prelude
 
 import Cardano.Api.UTxO (UTxO' (UTxO, toMap))
 import CardanoClient (
-  NodeLog,
   QueryPoint (QueryTip),
   RunningNode (..),
   buildAddress,
@@ -16,7 +15,7 @@ import CardanoClient (
   submitTx,
   waitForUTxO,
  )
-import CardanoNode (withCardanoNodeDevnet)
+import CardanoNode (NodeLog, withCardanoNodeDevnet)
 import Control.Concurrent.STM (newEmptyTMVarIO, takeTMVar)
 import Control.Concurrent.STM.TMVar (putTMVar)
 import Hydra.Cardano.Api (
